New Cyborg Generic Driver Proposal
Currently, the FPGA and GPU have been supported in Cyborg, the capability to
support the generic accelerator is not supported yet.

In general, these generic devices are the specific accelerator in some specific
scenario. For example:
1. The AI chips. which can be used for AI training and inference.
2. The security accelerator, which can be used for encryption and decryption.

In order to add the support for these specific accelerator, we propose to add
a generic driver to manage these devices. We propose to add a new driver,
"AcceleratorDriver", against exists FPGA, GPU driver.

Change-Id: I1efcb4e6bb51578c27b03dd8a800a5e289a94ff5