Specification of Cyborg API microversion
Many changes to the Cyborg REST API require changes to the consumers of
the API.
For example, if we need to add a required parameter to a method that is
called by Nova, we'd need both the Nova calling code and the
cyborgclient that Nova uses to change. But newer Cyborg versions with
the change must work with older Nova versions, and there is no mechanism
for this at the moment. Adding microversions will solve this problem.

APIImpact

Change-Id: I9b5f02daaf91ecd4416f032834c4fafdcb003b36